chore(shopify): delete legacy sync engine, jobs + ingress routes (phases 3-4)
Removes the backend half of the legacy Shopify customer/order/product sync,
now that the UI + dead tRPC surface are gone (#1004). The App-Store install,
billing provider/jobs, GDPR + billing webhooks, and chat-metafields are
untouched.

Lib / sync engine
- Delete shopify/sync-{customers,orders,products}.ts, shopify-initial-sync.ts,
  and the SyncManager (sync-manager.ts) — all legacy-sync-only.
- Delete jobs/shopify/** (sync + webhook jobs) and remove the worker
  registration (shopify-worker.ts + startShopifyWorker) and the shopifyQueue
  enum entry. Billing jobs (shopifyBillingSyncJob/shopifySeatUsageJob) stay.
- Delete contacts/sync-contact.ts (linkShopifyCustomer + its private
  syncContactFromSource helper — only the deleted customer sync used them).
- Delete shopify-webhooks.ts (setup/disable/create/deleteWebhook, autoSyncShopify).
  Relocate the one survivor, createShopifyAdminClient, to shopify/admin-client.ts
  (used by chat-metafields); repoint that import and slim shopify/index.ts.

Ingress routes
- Delete api/shopify/oauth2/** (legacy manual connect), api/shopify/webhook
  (sync dispatcher), and api/sync/[...slug] (legacy sync trigger) + the
  orphaned api/shopify/logger.ts.

Call-site cleanup
- user.ts: drop the autoSyncShopify call on the shopify.autoSync setting.
- seed organization-webhook-coordinator: drop the Shopify webhook
  disconnect/reconnect branches (keep the Google/Outlook email path).

Schema-table drops and the commerce seeder (table-coupled) come in phase 5.
