--- apps/web/src/app/api/shopify/logger.ts
@@ -1,3 +1,0 @@
-import { createScopedLogger } from '@auxx/logger'
-
-export const logger = createScopedLogger('api-shopify')
