+++ packages/lib/src/contacts/index.ts
@@ -3,4 +3,3 @@ export {
   ContactService,
   type ContactWithDetails,
 } from './contact-service'
-export { linkShopifyCustomer } from './sync-contact'
