--- packages/lib/src/jobs/shopify/index.ts
@@ -1,6 +1,0 @@
-export * from './customer-webhook-job'
-export * from './order-webhook-job'
-export * from './product-webhook-job'
-export * from './sync-customers-job'
-export * from './sync-orders-job'
-export * from './sync-products-job'
