+++ packages/lib/src/shopify/index.ts
@@ -1,9 +1,4 @@
-export * from '../sync-manager'
-export { SyncManager } from '../sync-manager'
+export { createShopifyAdminClient, type ShopifyAdminClient } from './admin-client'
 export * from './chat-metafields'
 export * from './shopify-types'
-export * from './shopify-webhooks'
-export * from './sync-customers'
-export * from './sync-orders'
-export * from './sync-products'
 export * from './utils'
